Remote procedure calls through CSharp(Sample).
Solution provides the following functionality for inter-process communication in .NET 5:

RPC using interface, particularly to replace WCF
RPC with asynchronous response, and
Streaming
The solution is based on SignalR library. It uses WebSocket as its main transport type with long polling as a fallback.

The following features are supported:

Binaries run in Windows and Linux alike
Communication channel may be encrypted with TLS.
Client is not necessarily .NET . It may be e. g. front end Web application.